chore: Remove various outdated files related to older Laravel version
This commit removes a substantial number of obsoleted files that were part of the old Laravel version. The deletion includes middleware files, providers, configuration files, command and exception files, and also language, migration, and event files. This clears out unnecessary files from the codebase, preparing it for the Laravel 11 upgrade.